Delete long running tests.
