--- s3/src/test/java/ch/cyberduck/core/sts/STSBucketRequestBeforeTokenExpiryFailsBecauseOfLatencyTest.java
@@ -1,84 +1,0 @@
-package ch.cyberduck.core.sts;
-
-/*
- * Copyright (c) 2002-2023 iterate GmbH. All rights reserved.
- * https://cyberduck.io/
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-import ch.cyberduck.core.Credentials;
-import ch.cyberduck.core.DisabledCancelCallback;
-import ch.cyberduck.core.DisabledHostKeyCallback;
-import ch.cyberduck.core.DisabledLoginCallback;
-import ch.cyberduck.core.Host;
-import ch.cyberduck.core.HostUrlProvider;
-import ch.cyberduck.core.OAuthTokens;
-import ch.cyberduck.core.Path;
-import ch.cyberduck.core.Profile;
-import ch.cyberduck.core.STSTokens;
-import ch.cyberduck.core.exception.BackgroundException;
-import ch.cyberduck.core.proxy.DisabledProxyFinder;
-import ch.cyberduck.core.s3.S3AccessControlListFeature;
-import ch.cyberduck.core.s3.S3FindFeature;
-import ch.cyberduck.core.s3.S3Session;
-import ch.cyberduck.test.TestcontainerTest;
-
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.testcontainers.containers.DockerComposeContainer;
-
-import java.util.EnumSet;
-
-import static ch.cyberduck.core.sts.AbstractAssumeRoleWithWebIdentityTest.MILLIS;
-import static ch.cyberduck.core.sts.AbstractAssumeRoleWithWebIdentityTest.OAUTH_TTL_SECS;
-import static ch.cyberduck.core.sts.STSTestSetup.*;
-import static org.junit.Assert.assertNotEquals;
-import static org.junit.Assert.assertTrue;
-
-@Category(TestcontainerTest.class)
-public class STSBucketRequestBeforeTokenExpiryFailsBecauseOfLatencyTest {
-
-    @ClassRule
-    public static DockerComposeContainer<?> compose = prepareDockerComposeContainer(
-            getKeyCloakFile()
-    );
-
-    @Test
-    @Ignore("Time of network latency may vary and so the time needs to be adjusted manually")
-    public void testBucketRequestBeforeTokenExpiryFailsBecauseOfLatency() throws BackgroundException, InterruptedException {
-        final Profile profile = readProfile();
-
-        final Host host = new Host(profile, profile.getDefaultHostname(), new Credentials("rawuser", "rawuser"));
-        final S3Session session = new S3Session(host);
-        session.open(new DisabledProxyFinder().find(new HostUrlProvider().get(host)), new DisabledHostKeyCallback(), new DisabledLoginCallback(), new DisabledCancelCallback());
-        session.login(new DisabledProxyFinder().find(new HostUrlProvider().get(host)), new DisabledLoginCallback(), new DisabledCancelCallback());
-
-        final OAuthTokens oauth = host.getCredentials().getOauth();
-        assertTrue(oauth.validate());
-        final STSTokens tokens = host.getCredentials().getTokens();
-        assertTrue(tokens.validate());
-
-        // Time of latency may vary and so the time needs to be adjusted accordingly
-        final int NETWORK_LATENCY = 1180;
-        final int wait = OAUTH_TTL_SECS * MILLIS - NETWORK_LATENCY;
-        Thread.sleep(wait);
-
-        Path container = new Path("cyberduckbucket", EnumSet.of(Path.Type.directory, Path.Type.volume));
-        assertTrue(new S3FindFeature(session, new S3AccessControlListFeature(session)).find(container));
-
-        assertNotEquals(oauth, host.getCredentials().getOauth());
-        assertNotEquals(tokens, host.getCredentials().getTokens());
-        session.close();
-    }
-}
