+++ s3/src/test/resources/testcontainer/keycloak/keycloak-realm.json
@@ -122,7 +122,7 @@
             "frontchannelLogout": false,
             "protocol": "openid-connect",
             "attributes": {
-                "access.token.lifespan": "30",
+                "access.token.lifespan": "5",
                 "saml.force.post.binding": "false",
                 "saml.multivalued.roles": "false",
                 "oauth2.device.authorization.grant.enabled": "false",
